# Parquet Reader: use aligned unpack in RleBpDecoder when possible