This significantly speeds up the `RleBpDecoder` when it contains long streaks of bitpacked data. 

###### Benchmark

```sql
SELECT UserID FROM hits WHERE UserID = 435090932899640449;
```

|  main  |  new   |
|--------|--------|
| 0.075s | 0.063s |